Use integer type instead of numeric

+++ src/composer/composeQuery.test.ts
@@ -127,7 +127,7 @@ describe("compose model queries", () => {
             operator: "is",
             rhs: {
               kind: "literal",
-              type: "numeric",
+              type: "integer",
               value: 1,
             },
           },
